Fix ccnt overflow in gpperfmon (#10205)
For some reason, gpmon_qexeckey_t structure used int16 for ccnt while all other GP code operates int32. This problem can cause ccnt overflow in gpperfmon packets.

This problem doesn't affect master branch as gpperfmon code has been removed from it.
But it seems to affect 6X_STABLE and 5X_STABLE branches.

Authored-by: Denis Smirnov darthunix@gmail.com

+++ src/include/gpmon/gpmon.h
@@ -192,7 +192,7 @@ typedef struct gpmon_qexec_hash_key_t {
 typedef struct gpmon_qexeckey_t {
     int32 tmid;  /* transaction time */
     int32 ssid; /* session id */
-    int16 ccnt;	/* command count */
+    int32 ccnt;	/* command count */
     gpmon_qexec_hash_key_t hash_key;
 }gpmon_qexeckey_t;
